requirements.txt
accelerate==1.12.0
asttokens==3.0.1
av==16.0.1
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
datasets==3.2.0
einops==0.8.1
executing==2.2.1
filelock==3.20.0
fsspec==2025.10.0
hf-xet==1.2.0
huggingface-hub==0.36.0
icecream==2.1.8
idna==3.11
Jinja2==3.1.6
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.5
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==11.3.0
psutil==7.1.3
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
qwen-vl-utils==0.0.14
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
six==1.17.0
sympy==1.13.1
tokenizers==0.22.1
torch==2.6.0+cu124
torchvision==0.21.0+cu124
tqdm==4.67.1
transformers==4.56.0
triton==3.2.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0